Extract chat discussion callbacks
