+++ service-worker.js
@@ -78,6 +78,7 @@ const APP_SHELL = [
   '/js/chat-nudge.js',
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
+  '/js/chat-discussion-callbacks.js',
   '/js/chat-discussion-round-request.js',
   '/js/chat-discussion-round-state.js',
   '/js/chat-discussion-round-view.js',
